requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pillow>=9.0.0
tqdm>=4.64.0
hydra-core>=1.3.0
omegaconf>=2.3.0
accelerate>=0.20.0
wandb>=0.15.0
lmdb>=1.4.0
webdataset>=0.2.31
scipy>=1.10.0
transformers>=4.30.0
deepspeed>=0.9.0
timm>=0.9.0
einops>=0.6.0
matplotlib>=3.7.0
click>=8.1.0
kagglehub>=0.2.0